third_party/WebKit/LayoutTests/external/wpt/editing/other/restoration-expected.txt

This is a testharness.js-based test.
PASS insertparagraph  starting bold
PASS insertparagraph  starting not bold
PASS insertlinebreak  starting bold
PASS insertlinebreak  starting not bold
FAIL delete  starting boldassert_equals: bold state expected false but got true
FAIL delete  starting not boldassert_equals: bold state expected true but got false
FAIL forwarddelete  starting boldassert_equals: bold state expected false but got true
FAIL forwarddelete  starting not boldassert_equals: bold state expected true but got false
FAIL insertorderedlist  starting boldassert_equals: bold state expected false but got true
FAIL insertorderedlist  starting not boldassert_equals: bold state expected true but got false
FAIL insertunorderedlist  starting boldassert_equals: bold state expected false but got true
FAIL insertunorderedlist  starting not boldassert_equals: bold state expected true but got false
FAIL indent  starting boldassert_equals: bold state expected false but got true
FAIL indent  starting not boldassert_equals: bold state expected true but got false
PASS outdent  starting bold
PASS outdent  starting not bold
FAIL justifyleft  starting boldassert_equals: bold state expected false but got true
FAIL justifyleft  starting not boldassert_equals: bold state expected true but got false
FAIL justifyright  starting boldassert_equals: bold state expected false but got true
FAIL justifyright  starting not boldassert_equals: bold state expected true but got false
FAIL justifycenter  starting boldassert_equals: bold state expected false but got true
FAIL justifycenter  starting not boldassert_equals: bold state expected true but got false
FAIL justifyfull  starting boldassert_equals: bold state expected false but got true
FAIL justifyfull  starting not boldassert_equals: bold state expected true but got false
FAIL formatblock div starting boldassert_equals: bold state expected false but got true
FAIL formatblock div starting not boldassert_equals: bold state expected true but got false
FAIL formatblock blockquote starting boldassert_equals: bold state expected false but got true
FAIL formatblock blockquote starting not boldassert_equals: bold state expected true but got false
PASS inserthorizontalrule  starting bold
FAIL inserthorizontalrule  starting not boldassert_equals: bold state expected true but got false
PASS insertimage a starting bold
FAIL insertimage a starting not boldassert_equals: bold state expected true but got false
PASS inserttext bar starting bold
PASS inserttext bar starting not bold
Harness: the test ran to completion.
